@@ -55,10 +55,10 @@
         "nav-frontend-veileder-style": "^3.0.1",
         "node-cache": "^5.1.2",
         "prop-types": "^15.8.1",
-        "react": "^17.0.2",
+        "react": "^18.2.0",
         "react-collapse": "^5.1.1",
         "react-dev-utils": "^12.0.1",
-        "react-dom": "^17.0.2",
+        "react-dom": "^18.2.0",
         "react-hot-loader": "^4.13.0",
         "react-refresh": "^0.14.0",
         "react-refresh-webpack-plugin": "^0.1.0",
@@ -91,8 +91,8 @@
         "@babel/runtime": "^7.18.6",
         "@types/fork-ts-checker-webpack-plugin": "^0.4.5",
         "@types/node": "^18.0.1",
-        "@types/react": "^17.0.39",
-        "@types/react-dom": "^18.0.5",
+        "@types/react": "^18.0.14",
+        "@types/react-dom": "^18.0.6",
         "@types/react-router-dom": "^5.3.3",
         "@types/styled-components": "^5.1.25",
         "@types/webpack": "^5.28.0",
@@ -130,5 +130,30 @@
         "webpack": "^5.73.0",
         "webpack-cli": "^4.10.0",
         "webpack-dev-server": "^4.9.3"
+    },
+    "overrides": {
+        "nav-frontend-chevron": {"react":  "$react" },
+        "nav-frontend-chevron-style": {"react":  "$react" },
+        "nav-frontend-core": {"react":  "$react" },
+        "nav-frontend-ikoner-assets": {"react":  "$react" },
+        "nav-frontend-js-utils": {"react":  "$react" },
+        "nav-frontend-lenker": {"react":  "$react" },
+        "nav-frontend-lenker-style": {"react":  "$react" },
+        "nav-frontend-lesmerpanel": {"react":  "$react" },
+        "nav-frontend-lesmerpanel-style": {"react":  "$react" },
+        "nav-frontend-lukknapp-style": {"react":  "$react" },
+        "nav-frontend-modal-style": {"react":  "$react" },
+        "nav-frontend-paneler": {"react":  "$react" },
+        "nav-frontend-paneler-style": {"react":  "$react" },
+        "nav-frontend-skjema": {"react":  "$react" },
+        "nav-frontend-skjema-style": {"react":  "$react" },
+        "nav-frontend-spinner": {"react":  "$react" },
+        "nav-frontend-spinner-style": {"react":  "$react" },
+        "nav-frontend-stegindikator": {"react":  "$react" },
+        "nav-frontend-stegindikator-style": {"react":  "$react" },
+        "nav-frontend-typografi": {"react":  "$react" },
+        "nav-frontend-typografi-style": {"react":  "$react" },
+        "nav-frontend-veileder": {"react":  "$react" },
+        "nav-frontend-veileder-style": {"react":  "$react" }
     }
 }
