@@ -1,15 +1,17 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
 import AppRoutes from './AppRoutes';
 import '@navikt/ds-css';
 import 'nav-frontend-core';
 import { AppProvider } from './context/AppContext';
+import { createRoot } from 'react-dom/client';
 
-ReactDOM.render(
+const container = document.getElementById('root');
+const root = createRoot(container!);
+
+root.render(
   <React.StrictMode>
     <AppProvider>
       <AppRoutes />
     </AppProvider>
-  </React.StrictMode>,
-  document.getElementById('root')
+  </React.StrictMode>
 );
